작업) 데이터 분리 작업
