+++ src/assets/api/about/aboutHotContentList.js
@@ -1,0 +1,24 @@
+const hotContentList = [
+  {
+    id: 1,
+    title: '퀴즈의 블랙홀',
+    img: '/images/quiz.png',
+  },
+  {
+    id: 2,
+    title: '위키드',
+    img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
+  },
+  {
+    id: 3,
+    title: '위키드',
+    img: '/images/wicked.png',
+  },
+  {
+    id: 4,
+    title: '위키드',
+    img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
+  },
+]
+
+export default hotContentList
