+++ src/components/about/AboutHotContents.jsx
@@ -3,31 +3,9 @@ import { Swiper, SwiperSlide } from 'swiper/react'
 import 'swiper/css'
 import 'swiper/css/pagination'
 import { Autoplay } from 'swiper/modules'
+import hotContentList from '../../assets/api/about/aboutHotContentList'
 
 function AboutHotContents() {
-  const contentList = [
-    {
-      id: 1,
-      title: '퀴즈의 블랙홀',
-      img: '/images/quiz.png',
-    },
-    {
-      id: 2,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 3,
-      title: '위키드',
-      img: '/images/wicked.png',
-    },
-    {
-      id: 4,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-  ]
-
   return (
     <>
       <AboutHotContentsWrap className="aboutHotContentsWrap">
@@ -44,7 +22,7 @@ function AboutHotContents() {
               disableOnInteraction: false,
             }}
             className="trending-swiper">
-            {contentList.map((item) => (
+            {hotContentList.map((item) => (
               <SwiperSlide key={item.id}>
                 <div className="content-card">
                   <div className="overlay"></div>
