+++ src/components/about/AboutUpdateContents.jsx
@@ -3,50 +3,9 @@ import { Swiper, SwiperSlide } from 'swiper/react'
 import 'swiper/css'
 import 'swiper/css/pagination'
 import { Pagination } from 'swiper/modules'
-// import { Autoplay } from 'swiper/modules'
+import updateContentList from '../../assets/api/about/aboutUpdateContentList'
 
 function AboutUpdateContents() {
-  const updateContentList = [
-    {
-      id: 1,
-      title: '나 혼자만 레벨 업',
-      desc: '10여 년 전, 다른 차원과 연결되는 게이트가 열리고 그 안에 숨어있는 마수들이 나타난다. 마수를 사냥하는 헌터들 중 랭크가 가장 낮은 최약체 성진우는 던전을 클리어하지만, 또 하나의 입구를 발견한다.',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 2,
-      title: '아노라',
-      desc: '뉴욕의 바에서 일하는 아노라는 러시아의 재벌 2세 이반을 만나 충동적으로 결혼식을 올린다. 신분상승을 꿈꾸던 것도 잠시, 이 사실을 알게 된 이반의 부모님은 미국에 있는 하수인을 시켜 혼인무효소송을 진행시킨다.',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 3,
-      title: '백설공주',
-      desc: '눈보라가 몰아치던 겨울 밤 태어난 백설공주. 온정이 넘치던 왕국에서 모두의 사랑을 받았지만, 강력한 어둠의 힘으로 왕국을 빼앗은 여왕의 위협에 숲으로 도망친다. 마법의 숲에서 간신히 살아남은 백설공주는 신비로운 일곱 광부들과 만나게 되며 새로운 세상을 마주하고, 마음속 깊이 숨겨진 용기와 선한 힘을 깨닫게 된다. 그리고 마침내, 빼앗긴 왕국을 되찾기 위해 여왕과 맞서 싸우기로 결심하는데…',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 4,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 5,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 6,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-    {
-      id: 7,
-      title: '위키드',
-      img: 'https://github.com/lse-7660/bokjak-image/blob/main/images/main/intro/introSlide1.png?raw=true',
-    },
-  ]
-
   return (
     <>
       <AboutUpdateContentsWrap className="aboutUpdateContentsWrap">
